+++ Adj/Program.hs
@@ -1,7 +1,5 @@
 module Adj.Program (module Exports) where
 
-import Adj.Program.Generation as Exports
-import Adj.Program.Maybe as Exports
-import Adj.Program.Progress as Exports
-import Adj.Program.Datastructure as Exports
+import Adj.Program.Primitive as Exports
 import Adj.Program.Controlflow as Exports
+import Adj.Program.Datastructure as Exports