+++ Adj/Program/Datastructure/Implementation/List.hs
@@ -1,7 +1,7 @@
 module Adj.Program.Datastructure.Implementation.List where
 
 import Adj.Auxiliary (type (=!?=))
-import Adj.Program.Generation (Construction)
-import Adj.Program.Maybe (Maybe)
+import Adj.Program.Primitive.Maybe (Maybe)
+import Adj.Program.Primitive.Generation (Construction)
 
 type List = Maybe =!?= Construction Maybe